Advance in getting full funcionality
After a lot of time I was able to resolve most of the challenges,
although now I'm starting to see a more clearer picture and I think
that I'll be able to improve the code. I was so deep in thought that
I completely forgot about committing often, so that's one thing to
have in mind in the next projects. Avoid this leaps from almost
nothing to almost everything.